\name{getW}
\alias{getW}
\title{
Calculating the weights of genes
}
\description{
This function gets the weights of nodes in \code{globalGraph}.
}
\usage{
getW(geneWeight, globalGraph)
}
\arguments{
  \item{geneWeight}{
The weights of genes.
}
  \item{globalGraph}{
The global pathway graph.
}
}
\value{
The weights of nodes in \code{globalGraph}.
}
